+++ eva/src/components/dashBoardInside/dashConstructorScheme/dashConstructorSchemesSettings.vue
@@ -27,7 +27,7 @@
               <template v-slot:item="{ item, on }">
                 <v-list-item
                   ripple
-                  class="v-list-item--link"
+                  class="v-list-item--link dash-constructor-schemes__data-panel-select"
                   v-on="on"
                 >
                   <v-list-item-content>
@@ -121,7 +121,7 @@
             <template v-slot:item="{ item, on }">
               <v-list-item
                 ripple
-                class="v-list-item--link"
+                class="v-list-item--link dash-constructor-schemes__data-panel-select"
                 v-on="on"
               >
                 <v-list-item-content>
@@ -158,7 +158,7 @@
             <template v-slot:item="{ item, on }">
               <v-list-item
                 ripple
-                class="v-list-item--link"
+                class="v-list-item--link dash-constructor-schemes__data-panel-select"
                 v-on="on"
               >
                 <v-list-item-content>
@@ -297,7 +297,7 @@
                 <template v-slot:item="{ item, on }">
                   <v-list-item
                     ripple
-                    class="v-list-item--link"
+                    class="v-list-item--link dash-constructor-schemes__data-panel-select"
                     v-on="on"
                   >
                     <v-list-item-content>
@@ -392,7 +392,7 @@
                 <template v-slot:item="{ item, on }">
                   <v-list-item
                     ripple
-                    class="v-list-item--link"
+                    class="v-list-item--link dash-constructor-schemes__data-panel-select"
                     v-on="on"
                   >
                     <v-list-item-content>
@@ -935,6 +935,10 @@ export default {
         color: var(--main-text);
       }
     }
+
+    ::v-deep.v-list-item--link {
+      max-width: 480px;
+    }
   }
   .dash-constructor-schemes__slider {
     .dash-constructor-schemes__slider-title {
@@ -943,5 +947,9 @@ export default {
       justify-content: space-between;
     }
   }
+}
+
+.dash-constructor-schemes__data-panel-select {
+  max-width: 480px;
 }
 </style>